# optimizer: add matvec optimizations