This PR still falls short of beating torch w/ mps.  The local cache will need to be enabled for that (as in the `extra/gemm/metal_matvec.py` example).

```
# with optimizations: python3 extra/gemm/metal_matvec.py
16384x4096    886.62 us, would be    151.38 GFLOPS matvec in torch
16384x4096    846.92 us, would be    158.48 GFLOPS matvec in metal
16384x4096    999.58 us, would be    134.27 GFLOPS matvec in tinygrad

# w/o optimizations: MV=0 python3 extra/gemm/metal_matvec.py
16384x4096    889.79 us, would be    150.84 GFLOPS matvec in torch
16384x4096    852.04 us, would be    157.52 GFLOPS matvec in metal
16384x4096   2066.08 us, would be     64.96 GFLOPS matvec in tinygrad
```

here's a long run with CNT=256 for `BIG=2 MPS=1 python3 test/test_speed_v_torch.py`:

<img width="1372" alt="Screenshot 2023-09-03 at 9 46 18 PM" src="https://github.com/tinygrad/tinygrad/assets/1693995/6ae461b8-ae90-4c9f-9e0d-864e303e18bf">