@tinyb0t 2023-09-04
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/codegen/linearizer.py             4               7
tinygrad/codegen/optimizer.py             27               0
------------------------------------------------------------
total                                     31               7
------------------------------------------------------------
lines added in the tinygrad folder: 24
```