@flammit 2023-09-04
Updated params and with `WINO=1 MV=1 BIG=2 MPS=1 python3 test/test_speed_v_torch.py`: 
<img width="1374" alt="Screenshot 2023-09-03 at 11 05 06 PM" src="https://github.com/tinygrad/tinygrad/assets/1693995/27265e63-014a-4246-b3fc-0dfcf9ba8c5f">