@geohot 2023-09-04
Assembly is removed, might make things easier