@flammit 2023-09-04
Small SD and larger JIT LLaMA increases (about 28%):

```
MV=1 PYTHONPATH="." python3 ./test/models/test_real_world.py                                   
test_gpt2: used 0.00 GB and 369 kernels in 129.88 ms
.test_llama: used 0.00 GB and 486 kernels in 49.39 ms
.test_sd: used 0.00 GB and 967 kernels in 243.24 ms
.train_cifar: used 0.00 GB and 153 kernels in 580.05 ms

MV=0 PYTHONPATH="." python3 ./test/models/test_real_world.py
test_gpt2: used 0.00 GB and 369 kernels in 129.77 ms
.test_llama: used 0.00 GB and 486 kernels in 68.73 ms
.test_sd: used 0.00 GB and 967 kernels in 251.60 ms
.train_cifar: used 0.00 GB and 153 kernels in 582.43 ms
```

Final performance numbers run with `WINO=1` and `CNT=1024` on M2 Max:

![Screenshot_2023-09-04_at_3 00 14_PM](https://github.com/tinygrad/tinygrad/assets/1693995/b8b01ca4-bcc3-4caa-93e7-7627e4c4a8ad)