@geohot 2023-09-06
Hmm, I'm on my boring M1 (not Max) now.

```
amanda@smol:~/tinygrad$ python3 extra/gemm/metal_matvec.py
16384x4096   4476.17 us, would be     29.98 GFLOPS matvec in torch
16384x4096   4997.75 us, would be     26.86 GFLOPS matvec in metal
16384x4096   5535.29 us, would be     24.25 GFLOPS matvec in tinygrad
amanda@smol:~/tinygrad$ MV=0 python3 extra/gemm/metal_matvec.py
16384x4096   4470.12 us, would be     30.03 GFLOPS matvec in torch
16384x4096   5066.92 us, would be     26.49 GFLOPS matvec in metal
16384x4096   4504.71 us, would be     29.79 GFLOPS matvec in tinygrad
amanda@smol:~/tinygrad$ 
```