@geohot 2023-09-06
I support the linearizer change regardless, we do really need ways to test things like that...I agree there's not a good framework yet